feat: reduce prettier print width to 100
I know this is a bit of a personal taste thing, but I've found a few times that I end up wondering why prettier isn't wrapping a very long line. If this were a lint rule, I'd want it to be even higher, but because prettier aims to fill the full available width, I think it would be better to reduce it slightly.