fix: 🐛 Improve eslint config to require fewer changes (#15)
Takes out the `no-underscore-dangle`, `camelcase`, and
`@typescript/naming-convention` rules as these don't match up with how
we were doing variable naming previously, Forbes concerns were correct,
and including these rules requires substantial changes. I also changed
`devDependencies` to allow using dev deps in test files, otherwise it
error on any use of dev dependencies with the `false` config the
conversion tool gave us. I also explicitly added the `jsx` parser
option. Also updated docs around eslint, globbing the files for eslint
is handled quite odd, and the `ext` option gets ignored. Instead the
better way to do it seems to just be pass it the `ext` options and give
it a directory. Also `eslint` didn't like a config within a package
after a proper install (I was using a local path to test originally), so
changed the docs to reflect this, requiring a `.eslintrc.js` to be made
in the project root directory.

These changes make the linter more permissive, so no breaking changes.

Also it's not used anywhere yet, so we wouldn't be breaking anything
even if it was.